feat(flamegraph): Add support for visualizing traces  (#1219)
* moved traceToFlamebeaver to flamegraph package and add trace property to FlamegraphRenderer
* add types for trace incomming data
* add LogoLink component and rename renderLogo to showCredit
* add proper flamegraph title
