+++ package.json
@@ -91,6 +91,8 @@
     "@types/lodash": "^4.14.176",
     "@types/lodash.debounce": "^4.0.6",
     "@types/lodash.defaults": "^4.2.6",
+    "@types/lodash.groupby": "^4.6.7",
+    "@types/lodash.map": "^4.6.13",
     "@types/mini-css-extract-plugin": "^2.4.0",
     "@types/node": "^17.0.7",
     "@types/prismjs": "^1.26.0",
@@ -202,6 +204,8 @@
     "lodash": "^4.17.21",
     "lodash.debounce": "^4.0.8",
     "lodash.defaults": "^4.2.0",
+    "lodash.groupby": "^4.6.0",
+    "lodash.map": "^4.6.0",
     "mini-css-extract-plugin": "^2.2.0",
     "moment": "^2.27.0",
     "msw": "^0.36.3",
