+++ packages/pyroscope-flamegraph/src/FlameGraph/FlameGraphComponent/ContextMenuHighlight.module.css
@@ -4,4 +4,5 @@
   /* make it a bit lighter so that it's easier to distinguish when both highlights are on */
   background: #ffffff8c;
   mix-blend-mode: overlay;
+  z-index: 2;
 }
