+++ packages/pyroscope-flamegraph/src/FlameGraph/FlameGraphComponent/canvas.module.css
@@ -1,4 +1,6 @@
 .canvas {
   width: 100%;
   cursor: pointer;
+  position: relative;
+  z-index: 1;
 }
