+++ packages/pyroscope-flamegraph/src/FlameGraph/FlameGraphComponent/styles.module.scss
@@ -2,3 +2,14 @@
   flex: 1;
   position: relative;
 }
+
+.logoLink {
+  text-decoration: none;
+  font-weight: 500;
+  opacity: 0.5;
+  color: var(--ps-neutral-2);
+  position: absolute;
+  right: 0;
+  bottom: 0;
+  z-index: 0;
+}
