+++ packages/pyroscope-models/src/index.ts
@@ -1,3 +1,4 @@
 export * from './profile';
 export * from './flamebearer';
+export * from './trace';
 export * from './decode';
