+++ webapp/javascript/pages/ContinuousComparisonView.tsx
@@ -114,6 +114,7 @@ function ComparisonApp() {
               }}
             />
             <FlamegraphRenderer
+              showCredit={false}
               panesOrientation="vertical"
               profile={leftSide}
               data-testid="flamegraph-renderer-left"
@@ -162,6 +163,7 @@ function ComparisonApp() {
               }}
             />
             <FlamegraphRenderer
+              showCredit={false}
               profile={rightSide}
               data-testid="flamegraph-renderer-right"
               panesOrientation="vertical"
