+++ webapp/javascript/pages/ContinuousSingleView.tsx
@@ -55,6 +55,7 @@ function ContinuousSingleView() {
       case 'reloading': {
         return (
           <FlamegraphRenderer
+            showCredit={false}
             profile={singleView.profile}
             colorMode={colorMode}
             ExportData={
